ThemePicker: Restore wallpaper picker
When theming components were removed in a previous commit, the ability to
change the wallpaper was inadvertently removed as well, so bring it back.

Note that the restored wallpaper picker looks a little funky without the
theme color picker, as both combined had a complete rounded rectangle. This
will be rectified in a future commit.